refactor: atualiza os nomes dos arquivos com sufixo e as importacoes
